% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_data.R
\docType{data}
\name{agCensusAcres}
\alias{agCensusAcres}
\title{agCensusAcres}
\format{
A data frame with 15369 rows and 15 columns covering Inf--Inf.
}
\source{
USDA NASS Quick Stats
}
\usage{
data(agCensusAcres)
}
\description{
A combined dataset for agCensusAcres
}
\keyword{datasets}
