% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_data.R
\docType{data}
\name{agCensusBFR}
\alias{agCensusBFR}
\title{agCensusBFR}
\format{
A data frame with 255 rows and 16 columns covering NA-NA.
}
\source{
USDA NASS Quick Stats
}
\usage{
data(agCensusBFR)
}
\description{
A combined dataset for agCensusBFR
}
\keyword{datasets}
