% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_data.R
\docType{data}
\name{elasticity_limits}
\alias{elasticity_limits}
\title{Simulator Helper Datasets}
\format{
A data frame with 1 rows and 4 columns covering Inf--Inf.
}
\source{
Elasticity limits determined as 50\% of the magnitude of the program level elasticities
}
\usage{
data(elasticity_limits)
}
\description{
A combined dataset for elasticity_limits
}
\keyword{datasets}
