% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_data.R
\docType{data}
\name{fcip_contiguous_county}
\alias{fcip_contiguous_county}
\title{fcip_contiguous_county}
\format{
A data frame with 24307 rows and 12 columns covering Inf--Inf.
}
\source{
USDA-RMA, Actuarial Data Master - A0123
}
\usage{
data(fcip_contiguous_county)
}
\description{
A combined dataset for fcip_contiguous_county
}
\keyword{datasets}
