% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_data.R
\docType{data}
\name{fcip_recodes_commodity_groupings}
\alias{fcip_recodes_commodity_groupings}
\title{fcip_recodes_commodity_groupings}
\format{
A data frame with 3668 rows and 10 columns covering 1997-2026.
}
\source{
USDA-RMA, Actuarial Data Master - A00400 and A00420 supplemented data from legacy ADM files
}
\usage{
data(fcip_recodes_commodity_groupings)
}
\description{
A combined dataset for fcip_recodes_commodity_groupings
}
\keyword{datasets}
