% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_data.R
\docType{data}
\name{fcip_recodes_insurance_plan}
\alias{fcip_recodes_insurance_plan}
\title{fcip_recodes_insurance_plan}
\format{
A data frame with 773 rows and 10 columns covering 1989-2025.
}
\source{
USDA-RMA, Actuarial Data Master - A00460 supplemented data from legacy ADM files
}
\usage{
data(fcip_recodes_insurance_plan)
}
\description{
A combined dataset for fcip_recodes_insurance_plan
}
\keyword{datasets}
