% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_data.R
\docType{data}
\name{fcip_recodes_practice}
\alias{fcip_recodes_practice}
\title{fcip_recodes_practice}
\format{
A data frame with 29562 rows and 8 columns covering 1997-2026.
}
\source{
USDA-RMA, Actuarial Data Master - A00510 supplemented data from legacy ADM files
}
\usage{
data(fcip_recodes_practice)
}
\description{
A combined dataset for fcip_recodes_practice
}
\keyword{datasets}
