% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_data.R
\docType{data}
\name{fcip_recodes_type}
\alias{fcip_recodes_type}
\title{fcip_recodes_type}
\format{
A data frame with 232730 rows and 7 columns covering 1999-2025.
}
\source{
Generated internally, using harmonize_crop_type_codes()
}
\usage{
data(fcip_recodes_type)
}
\description{
A combined dataset for fcip_recodes_type
}
\keyword{datasets}
