% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_data.R
\docType{data}
\name{fsa_crop_linker}
\alias{fsa_crop_linker}
\title{Simulator Helper Datasets}
\format{
A data frame with 4389 rows and 8 columns covering Inf--Inf.
}
\source{
Internal innovation
}
\usage{
data(fsa_crop_linker)
}
\description{
A combined dataset for fsa_crop_linker
}
\keyword{datasets}
