% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global_variables.R
\docType{data}
\name{global_variables}
\alias{global_variables}
\title{fsa_crop_linker}
\format{
A data frame with 8594 rows and 8 columns covering Inf--Inf.
}
\source{
Internal innovation
}
\usage{
global_variables
}
\description{
A combined dataset for global_variables
}
\keyword{datasets}
