% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_data.R
\docType{data}
\name{nass_census_state_beginning_farmer_and_rancher_data}
\alias{nass_census_state_beginning_farmer_and_rancher_data}
\title{nass_census_state_beginning_farmer_and_rancher_data}
\format{
A data frame with 255 rows and 16 columns covering Inf--Inf.
}
\source{
USDA NASS Quick Stats
}
\usage{
data(nass_census_state_beginning_farmer_and_rancher_data)
}
\description{
A combined dataset for nass_census_state_beginning_farmer_and_rancher_data
}
\keyword{datasets}
