% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_data.R
\docType{data}
\name{nass_index_for_price_recived}
\alias{nass_index_for_price_recived}
\title{nass_index_for_price_recived}
\format{
A data frame with 35 rows and 3 columns covering 1990-2024.
}
\source{
USDA NASS Quick Stats
}
\usage{
data(nass_index_for_price_recived)
}
\description{
A combined dataset for nass_index_for_price_recived
}
\keyword{datasets}
