% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_data.R
\docType{data}
\name{nass_marketing_year_avg_price}
\alias{nass_marketing_year_avg_price}
\title{nass_marketing_year_avg_price}
\format{
A data frame with 31145 rows and 7 columns covering 1866-2024.
}
\source{
USDA NASS Quick Stats
}
\usage{
data(nass_marketing_year_avg_price)
}
\description{
A combined dataset for nass_marketing_year_avg_price
}
\keyword{datasets}
