% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_data.R
\docType{data}
\name{nass_state_rental_rates}
\alias{nass_state_rental_rates}
\title{nass_state_rental_rates}
\format{
A data frame with 1792 rows and 5 columns covering 1994-2025.
}
\source{
Output from get_state_rental_rates() function
}
\usage{
data(nass_state_rental_rates)
}
\description{
A combined dataset for nass_state_rental_rates
}
\keyword{datasets}
