% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_data.R
\docType{data}
\name{nass_us_ag_price_index_monthly}
\alias{nass_us_ag_price_index_monthly}
\title{nass_us_ag_price_index_monthly}
\format{
A data frame with 2799 rows and 8 columns covering Inf--Inf.
}
\source{
USDA NASS: https://www.nass.usda.gov/Charts_and_Maps/graphics/data
}
\usage{
data(nass_us_ag_price_index_monthly)
}
\description{
A combined dataset for nass_us_ag_price_index_monthly
}
\keyword{datasets}
