% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_data.R
\docType{data}
\name{nassAgPriceMonthlyIndex}
\alias{nassAgPriceMonthlyIndex}
\title{nassAgPriceMonthlyIndex}
\format{
A data frame with 2625 rows and 8 columns covering Inf--Inf.
}
\source{
USDA NASS: https://www.nass.usda.gov/Charts_and_Maps/graphics/data
}
\usage{
data(nassAgPriceMonthlyIndex)
}
\description{
A combined dataset for nassAgPriceMonthlyIndex
}
\keyword{datasets}
