% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_data.R
\docType{data}
\name{nassSeasonLengthState}
\alias{nassSeasonLengthState}
\title{nassSeasonLengthState}
\format{
A data frame with 9523 rows and 14 columns covering 1981-2025.
}
\source{
Calculated
}
\usage{
data(nassSeasonLengthState)
}
\description{
A combined dataset for nassSeasonLengthState
}
\keyword{datasets}
