% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_data.R
\docType{data}
\name{nassSurveyInputApplications}
\alias{nassSurveyInputApplications}
\title{nassSurveyInputApplications}
\format{
A data frame with 3830 rows and 25 columns covering 1990-2024.
}
\source{
USDA NASS Quick Stats
}
\usage{
data(nassSurveyInputApplications)
}
\description{
A combined dataset for nassSurveyInputApplications
}
\keyword{datasets}
