% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_data.R
\docType{data}
\name{nassSurveyMYAprice}
\alias{nassSurveyMYAprice}
\title{nassSurveyMYAprice}
\format{
A data frame with 33078 rows and 7 columns covering 1866-2024.
}
\source{
USDA NASS Quick Stats
}
\usage{
data(nassSurveyMYAprice)
}
\description{
A combined dataset for nassSurveyMYAprice
}
\keyword{datasets}
