% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_data.R
\docType{data}
\name{nassSurveyPriceRecivedIndex}
\alias{nassSurveyPriceRecivedIndex}
\title{nassSurveyPriceRecivedIndex}
\format{
A data frame with 36 rows and 3 columns covering 1990-2025.
}
\source{
USDA NASS Quick Stats
}
\usage{
data(nassSurveyPriceRecivedIndex)
}
\description{
A combined dataset for nassSurveyPriceRecivedIndex
}
\keyword{datasets}
