% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_data.R
\docType{data}
\name{nassSurveyRentalRates}
\alias{nassSurveyRentalRates}
\title{nassSurveyRentalRates}
\format{
A data frame with 1792 rows and 5 columns covering 1994-2025.
}
\source{
Output from get_state_rental_rates() function
}
\usage{
data(nassSurveyRentalRates)
}
\description{
A combined dataset for nassSurveyRentalRates
}
\keyword{datasets}
