% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_data.R
\docType{data}
\name{premium_subsidy_schedule}
\alias{premium_subsidy_schedule}
\title{premium_subsidy_schedule}
\format{
A data frame with 5528 rows and 7 columns covering 2010-2025.
}
\source{
USDA-RMA, Actuarial Data Master supplemented data from legacy ADM files
}
\usage{
data(premium_subsidy_schedule)
}
\description{
A combined dataset for premium_subsidy_schedule
}
\keyword{datasets}
