% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_data.R
\docType{data}
\name{prevented_planting_adm}
\alias{prevented_planting_adm}
\title{prevented_planting_adm}
\format{
A data frame with 252605 rows and 10 columns covering 2011-2026.
}
\source{
USDA-RMA, Actuarial Data Master supplemented data from legacy ADM files
}
\usage{
data(prevented_planting_adm)
}
\description{
A combined dataset for prevented_planting_adm
}
\keyword{datasets}
